+++ src/slashCommand.ts
@@ -57,8 +57,9 @@ export class SlashCommandManager {
 
   static async create(context: Context) {
     const pluginSlashCommands = await context.apply({
-      hook: 'command',
+      hook: 'slashCommand',
       args: [],
+      memo: [],
       type: PluginHookType.SeriesMerge,
     });
     return new SlashCommandManager({